// src/components/feature/Main/RegisterUser/interface.ts
type InputType =
  | "button"
  | "checkbox"
  | "color"
  | "date"
  | "datetime-local"
  | "email"
  | "file"
  | "hidden"
  | "image"
  | "month"
  | "number"
  | "password"
  | "radio"
  | "range"
  | "reset"
  | "search"
  | "submit"
  | "tel"
  | "text"
  | "time"
  | "url"
  | "week";

export interface configRegElem {
  name: string;
  forId: string;
  text: string;
  type: InputType;
  placeholder: string;
  value?: string;
}